+++ src/core/transaction/index.ts
@@ -1,6 +1,7 @@
 export * from './step';
 export * from './mapping';
 export * from './replace-step';
+export * from './replace-document-step';
 export * from './replace-node-step';
 export * from './insert-text-step';
 export * from './replace-text-step'